exo piscine

+++ piscine/fibo_iter/fibo_iter.c
@@ -1,15 +1,15 @@
-#include <stdio.h>
 #include <stddef.h>
+#include <stdio.h>
 
-unsigned long fibo_iter(unsigned long n){
-    
+unsigned long fibo_iter(unsigned long n)
+{
     unsigned long n1 = 0;
     unsigned long n2 = 1;
-    
+
     if (n == 0)
         return 0;
 
-    for (int i = 1; i<n; i++)
+    for (unsigned long i = 1; i < n; i++)
     {
         unsigned long temp;
         temp = n2;
